requirements.txt
aiohttp==1.0.5
astroid==2.1.0
async-timeout==3.0.1
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
discode==0.0.9
discord.py==0.16.12
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.3.0
mccabe==0.6.1
multidict==4.5.2
pylint==2.2.2
requests==2.21.0
six==1.12.0
soupsieve==1.7.2
typed-ast==1.2.0
urllib3==1.24.1
websockets==3.4
wrapt==1.11.1
